feat: separate GET usage
Not all getUsed is equal, and these numbers cannot be combined. On
minting an NFT this value represents GET moving from the silo to the NFT
fuel tank. Upon check-in this would be from the fuel tank to the depot.

As a result these have been separated out into individual counts.

Counts have also been returned to the Event entity.