feat: add Tickets entity
Lots of use-cases for this one that it felt a worthwhile addition before
we kick off the full index. For example:

- Recreate the ticket explorer lifecycle page only using the subgraph.
- See the fuel tank balance of the NFT (can use to discover which have
  large remaining balances)
- Offer direct query data to the user per-ticket.
- See the tickets for an event.

There are a few oddities to this one worth calling out:

- On v1 of the contracts we were submitting the primaryPrice as
  denominated in the local currency. This has now been approximated
  using an averge USD rate for each of the active currencies. From now
  on the basePrice is available in USD and requries no static
  conversion.
- Many of the missing EventMetadataStorage fields have also been added.

Due to storing the Ticket with the eventAddress we save a lot of queries
to the blockchain node. Previously the eventAddress was required for
nearly every action but now that this is cached on the Ticket entity we
reduce the load on the node significantly.

This is unfortunately still required for V2 mints (as we only have the
nftIndex for efficiency) but under the batching of V3 we'll likely end
up in a better state after V2. It does however save queries to the
blockchain node for all actions aside from PrimarySaleMint under V2.