fix: uint64 rounding correction
The gasUsed argument on each emitted event on the v2 contracts is a
uint64 cast from a source of uint256, causing some inaccuracy. This
results in 'rounding down' to the floor of the uint64 precision for each
ticket minted causing a slight drift in the aggregated values over time.
For obvious reasons the gasUsed needs to be 100% accurate and will be
emitted as a uint256 in future versions of the contract.

The workaround here is to fetch the fuel tank balance of the NFT after
minting as this is returned as a uint256. Scanning, checking-in, and
invalidating all clear the fuel tank, so the existing fuel tank balance
can be used as the value of gasUsed for these actions.

See v2 README.md comment for more.