fix: add ABI to correct consumer
All ABIs used from an event issued by a contract must be listed against
that contract. Somewhat obvious in hindsight but had expected that it
would search the yaml for the appropriate contract you're binding to and
then use its ABI - not the case.

If you wish to bind to EconomicsGET from an event issued by BaseGETV2
then you must list this ABI under BaseGETV2, for instance.