feat: add economics events
Broad PR to add economics events (primarily top-ups) to the subgraph.
This allows us to track the siloBalance, depotBalance, and basePrice as
a real-time measure as well as the historical end-of-day value.

The ProtocolDay entity will contain the depotBalance with the last
update for that 24 hour period, wheras Protocol will contain the
depotBalance to the latest update. Likewise for the RelayerDay and
Relayer entities. This allows for historical tracking/graphing/charting
while still offering a real-time headline measurement per-relayer.

While here, also;
- Updated the ABIs to the latest as found on Polygonscan.
- ERD and docs updated to latest models.