feat: v2
Bumper changeset to migrate-forward existing data into the v2 schema as
well as index the new data on the v2 contracts. A lot has changed here,
here are the main ones.

While many of the existing entities on v1 will be recognisable, there
are some decisions of the past that we have chosen to rotate away from.
Some of these are terminology related, others are to do with shared
accounting. The devil is somewhat in the details with these and we
sometimes need to make some judgement calls on the best way to migrate
this data forwards.

**1. New Schema**

Accounting using a shared relayer is a key example of this, because at
present many ticketeers share a single relayer. Under v2 this is not
going to the be the case and each integrator will have their own
independent accounting and relayer addresses. But the tickets already
minted have used the shared relayer, how is this worked around? Two
fold; for existing tickets we split the accounting responsibilities from
the event data ownership. For this we leverage `Event.ticketeerName` to
assign the integrator that 'owns' the event (who created it and ran it),
from the integrator account that provided the fuel, which in this case
will be a shared billing account. These mappings are provuded in
`RELAYER_MAPPING` and `TICKETEER_MAPPING` within
src/constants/contracts.ts.

**3. New Entities**

A couple of new entities have popped up; Integrator, IntegratorDay,
SpentFuelRecipient, SpentFuelCollectedEvent, and PriceOracle. The
purpose of these have been documented in the README.md.

**3. Terminology Simplification**

RIP Silo, RIP Depot, we hardly knew ye.

After some discussion we decided to reduce parts of the fuel lifecyle
metaphor that was present in the codebase and documentation, in practice
this means that:

- The 'siloBalance' becomes the 'availableFuel'. This is the balance
  available to be used to pay for NFT ticketing functionality.
- Reserved fuel is what is available in the system. Previously the fuel
  tanks. For efficiency this is now an integrator-wide measurement and
  no longer per-ticket.
- Spent fuel is the new depot balance. This is the amount of reserved
  fuel that has been used as is available for collection. Typically
  moved from reserved to spent on check-in and invalidations.

"Available" (to be spent)
  -> "Reserved" (held in the system)
    -> "Spent" (awaiting collection)