feat: production fixes and config
Running through a number of record association fixes on the production
data set, in particular linking relayers using their address.

This commit also includes the real development configurations as well as
separating out the v1 subgraph using grafts. This means that on the main
subgraph deployment we only need to index from when the new contracts
were released and not from the beginning of time, making it much faster
to re-index all events on any change.