fix: is<x> flag not storing on V1 events
Due to not calling `ticket.save()` after setting the variable locally,
this wasn't being saved across the board. This commit ensures the ticket
is saved on all events.

While here (and needing a fresh re-index) the `totalTicketValue` field
has been renamed to `totalSalesVolume` to make this more clear that it
represents sales and not a more arbitrary form of value measurement.