fix: divide-by-zero on secondary sale
Nasty bug wherein a resold ticket being the first action of the day can
brick the subgraph. This is because to calculate the average GET fuel
per-ticket we must divide by the number of tickets minted that day,
which can be zero.

This fixes the handling so that the average fuel per-day is only set
when the sold count for that day is greater than 0.