feat: calculate fuel per-ticket
Previously the fuel used by a ticket in a batch was the total fuel used
in the batch divided by the number of tickets. This meant that tickets
of varying basePrice would have the same fuel usage when batched
together.

Because the events do not emit the fuel used per-ticket we instead
calculate the fuel usage using the same computation as on the contracts.
This gives us an accurate value for Ticket.reservedFuel, albeit at the
expense of duplication of code.