feat: decentralized service
Launch the subgraphs on the decentralized service. To facilitate
reproducible builds we also introduce a Dockerfile that managed
dependencies. Building and releasing the subgraph to the decentralized
service can then be done by:

docker build --build-arg DEPLOY_KEY=<KEY> -t subgraph .
docker run -it subgraph release:<ENV> -l <VERSION>

Although not covered in this PR it is expected that these commands can
be re-used in a CI environment such as Github Actions for automated
releases.

Also implemented:

- Update packages.
- Update graft base ID.