fix: handle empty ticket actions on action events
Prior to now there was no need for this fix because any action
within a batch that didn't pass the action rules would have the
batch reverted. This was changed in favour of dropping those
batches and emitting them as ActionErrorLogs.
This has made it possible for all actions to be dropped and have
an action event emitted with an empty array of ticket actions.
When this happen, we just ignore that event.